requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlite3-to-pandas==0.3.0
sqlalchemy==2.0.23
alembic==1.12.1

# AI/ML Libraries
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langgraph==0.0.19
langsmith==0.0.69
openai==1.3.8

# HTTP Requests & APIs
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.1

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Async & Concurrency
asyncio-throttle==1.0.2

# Logging & Monitoring
structlog==23.2.0

# Web Scraping (for research)
beautifulsoup4==4.12.2
selenium==4.16.0

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# WebSocket Support (for real-time dashboard)
websockets==12.0

# CORS
fastapi-cors==0.0.6

# JSON Schema validation
jsonschema==4.20.0